pyproject.toml
[project]
name = "wschat"
version = "0.1.0"
description = ""
authors = [
    {name = "Zero Two",email = "hau.au.999@gmail.com"}
]
readme = "README.md"
# Такой странный паттерн - для установки wemake-python-styleguide
requires-python = "<4.0,>=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "argon2-cffi (>=23.1.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
wemake-python-styleguide = "^1.1.0"

